requirements.txt
Django==2.2.2
django-filter==2.1.0
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
Markdown==3.1.1
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
